requirements.txt
mysql-connector-python==8.0.32
environs~=9.5.0
aiogram~=3.0.0b7